+++ CHANGELOG.md
@@ -10,7 +10,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * larva-css - Update css generator script to include desktop_xl breakpoint using media queries
 * larva-css - Update test for generated css
 * larva-patterns - Add `author` and `author-details` module.
+* larva-js - Add popover js
 * larva-tokens - Update grey color for rollingstone-2022.
+* larva-patterns - Add `article-title` module.
+* larva-patterns - Add `article-kicker` module.
+* larva-patterns - Add `article-excerpt` module.
 
 ## 0.5.5 - 03-03-2022
 * larva-css - Add border top utility.